+++ noodles-bcf/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Changed
 
+  * bcf/async/io/reader/header: Parse header line by line.
+
+    The async header reader now builds a `vcf::Header` by parsing a raw header
+    line by line. This makes it so that it is no longer required to read the
+    entire raw header into memory before parsing.
+
   * bcf/async/io/reader/header: Discard VCF header padding.
 
 ## 0.66.0 - 2024-12-20